add oc common
